requirements.txt
paho-mqtt==1.4.0
speedtest-cli==2.0.2
geopy==1.18.1
schedule==0.6.0
datetime==4.3